Add files via upload
Added functionality for a timeline view that sorts in alphabetical order. I had a bug with css for buttons where hover wouldn't override their width, as ideally in larger timelines users could mouse over an item and it would make it more visible. My solution to this was just making the text a lot bigger, which gets the job done but looks a little ugly. If there's a way to override the width I doubt it would take a lot of changes. The timeline only sorts in alphabetical order but I'd definitely be willing to add a menu that allows for sorting by different criteria, like with the schema class. Also, the buttons currently don't return anything on click, but I can also have it return some data about the nodes.