Added a basic set of Api utilities for scraping tables
Added ApiUtils which contains a couple utilities that help with scraping
arbitrary tables from webpages and converting them into collection documents.
Also added basic test UI that allows entering a url and arbitrary CSS
selector to select the table from the page. You can then query the columns of
the table and choose which ones you want to be included in the final document,
and then you can create the collection and view it in a (poorly formatted)
schema view.