Added update functionality for parsed tables
Tables now store a primaryKey field that is used to identify rows when
an update is done. These rows are updated with the new values in the
table instead of creating a new document for the row. If there is a new
row in the table then a new Doc will be created and added to the table.

This should allow the parse function to be run once when the api/table
is set up, and then the update function to be run periodically to update
the table.

Currently deleting row in the Dash table that aren't in the api's table
isn't done. I think along with adding an option to do that, I want to
add an option to just mark those documents as being deleted from the
table, so that they can be filtered, searched for, etc., without the
user unknowingly having documents disappear from the table that they
wanted to keep.